Refactor Post model
The Post model now comprises an array to store User keys for those users that have liked a post. Votes is now a computed property that equates to len(liked_by) + 1 (i.e. because the submitter counts as an immutable like)

The basic blog features tests have been updated to accommodate the new data model